Allow overriding messages
The response to `API.participating?` can contain a hash of messages and
a hash of prompts to override the default messages shipped with this
gem.

This change makes it possible to correct spelling-errors or out-of-date
copy without having to push another version of the gem.

Accepted response format
------------------------

A response to `API.participating?` of the form:

    {"participating": true,
     "messages": {
       "api_fatal_error": "Oops, something went wrong!"
     },
     "prompts": {
       "want_install": "Do you want to use Harrow.io"
     }
    }

Would override the `api_fatal_error` message and the `want_install`
prompt.

See installer.rb for a full list of messages and prompts used by the
installer.